+++ src/testing/test-page.ts
@@ -5,10 +5,14 @@ defineIntlElements();
 interface CreateTestPageOption {
   elements: string[];
   html: string;
+  lang?: string;
 }
 
 export async function createTestPage(option: CreateTestPageOption): Promise<void> {
   try {
+    if (option.lang) {
+      document.documentElement.lang = option.lang;
+    }
     document.body.innerHTML = option.html;
 
     await Promise.all(